requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1 
Flask-HTTPAuth==4.5.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
Flask-Migrate== 4.0.5 
python-dotenv == 1.0.0
aniso8601==9.0.1
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
pytz==2021.3
six==1.16.0
Werkzeug==3.0.2
pandas==2.2.2
webargs==8.4.0





